MAVLink PR150 rebased
* Merged EdgeTX PR EdgeTX#150
* Added MAVLink YAML elements.
* Set fastmavlink to be8b000 and mavlink to 37a6e90.

+++ radio/src/telemetry/telemetry.h
@@ -44,7 +44,7 @@
 
 #include "pulses/modules_helpers.h"
 
-extern uint8_t telemetryStreaming; // >0 (true) == data is streaming in. 0 = no data detected for some time
+extern uint16_t telemetryStreaming; // >0 (true) == data is streaming in. 0 = no data detected for some time
 
 inline bool TELEMETRY_STREAMING()
 {
